requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data processing
pandas==2.1.3
numpy==1.26.2
duckdb==0.9.2

# Social media APIs
praw==7.7.1
tweepy==4.14.0

# NLP and sentiment analysis
vaderSentiment==3.3.2
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99

# Stock data
polygon-api-client==1.13.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
plotly==5.18.0
